criação duas pastas backend e frontend - instalação react

+++ .docker/entrypoint.sh
@@ -1,8 +1,23 @@
 #!/bin/bash
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
-cp .env.example .env
-cp .env.testing.example .env.testing
+
+
+## FRONTEND ##
+npm config set cache /var/www/.npm-cache --global
+cd /var/www/frontend && npm install && cd ..
+
+
+## BACKEND ##
+
+cd backend
+
+if [ ! -f ".env" ]; then
+  cp .env.example .env
+fi
+if [ ! -f ".env.testing" ]; then
+  cp .env.testing.example .env.testing
+fi
 php artisan key:generate
 composer install
 php artisan migrate
